Refactors getting data for charts, renames file.
I _think_ this doesn't break anything. Trying to break this up
into several commits.
